pyproject.toml
[project]
name = "multi-agent-task-system"
version = "1.0.0"
description = "A personal productivity orchestrator — AI agents that manage your calendar and tasks via MCP"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # --- Agent Framework ---
    "google-adk>=1.0.0",

    # --- API Server ---
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",

    # --- Database ---
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",

    # --- Google APIs ---
    "google-api-python-client>=2.140.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",

    # --- Notion API ---
    "notion-client>=2.2.0",

    # --- MCP ---
    "mcp>=1.0.0",

    # --- Scheduling (Tier 3) ---
    "apscheduler>=3.10.0",

    # --- Config ---
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",

    # --- HTTP Client ---
    "httpx>=0.27.0",

    # --- Vertex AI fallback ---
    "google-cloud-aiplatform>=1.60.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
